move list area normal back a level
