+++ shared/chat/conversation/list-area/index.native.tsx
@@ -1,16 +1,16 @@
-import * as Container from '../../../../util/container'
-import * as Kb from '../../../../common-adapters/mobile.native'
+import * as Container from '../../../util/container'
+import * as Kb from '../../../common-adapters/mobile.native'
 import * as React from 'react'
-import * as Styles from '../../../../styles'
-import * as Types from '../../../../constants/types/chat2'
+import * as Styles from '../../../styles'
+import * as Types from '../../../constants/types/chat2'
 import JumpToRecent from './jump-to-recent'
-import Message from '../../messages'
-import SpecialBottomMessage from '../../messages/special-bottom-message'
-import SpecialTopMessage from '../../messages/special-top-message'
-import logger from '../../../../logger'
+import Message from '../messages'
+import SpecialBottomMessage from '../messages/special-bottom-message'
+import SpecialTopMessage from '../messages/special-top-message'
+import logger from '../../../logger'
 import {Animated, ListRenderItemInfo} from 'react-native'
 import {Props, ItemType} from '.'
-import {mobileTypingContainerHeight} from '../../input-area/normal/typing'
+import {mobileTypingContainerHeight} from '../input-area/normal/typing'
 
 const debugEnabled = false
