+++ shared/chat/conversation/list-area/index.stories.tsx
@@ -1,20 +1,20 @@
 /* eslint-disable sort-keys */
 import * as React from 'react'
-import * as Sb from '../../../../stories/storybook'
+import * as Sb from '../../../stories/storybook'
 import moment from 'moment'
-import {Button, ButtonBar, Box2, Text} from '../../../../common-adapters'
-import * as Types from '../../../../constants/types/chat2'
-import {propProvider as ReactionsRowProvider} from '../../messages/reactions-row/index.stories'
-import {propProvider as ReactButtonProvider} from '../../messages/react-button/index.stories'
-import {propProvider as ReactionTooltipProvider} from '../../messages/reaction-tooltip/index.stories'
-import {OwnProps as ExplodingMetaOwnProps} from '../../messages/wrapper/exploding-meta/container'
-import {Props as ExplodingMetaViewProps} from '../../messages/wrapper/exploding-meta'
+import {Button, ButtonBar, Box2, Text} from '../../../common-adapters'
+import * as Types from '../../../constants/types/chat2'
+import {propProvider as ReactionsRowProvider} from '../messages/reactions-row/index.stories'
+import {propProvider as ReactButtonProvider} from '../messages/react-button/index.stories'
+import {propProvider as ReactionTooltipProvider} from '../messages/reaction-tooltip/index.stories'
+import {OwnProps as ExplodingMetaOwnProps} from '../messages/wrapper/exploding-meta/container'
+import {Props as ExplodingMetaViewProps} from '../messages/wrapper/exploding-meta'
 import Thread from '.'
-import * as Message from '../../../../constants/chat2/message'
-import HiddenString from '../../../../util/hidden-string'
+import * as Message from '../../../constants/chat2/message'
+import HiddenString from '../../../util/hidden-string'
 import JumpToRecent from './jump-to-recent'
-import SpecialTopMessage from '../../messages/special-top-message'
-import * as Constants from '../../../../constants/chat2'
+import SpecialTopMessage from '../messages/special-top-message'
+import * as Constants from '../../../constants/chat2'
 
 const firstOrdinal = 10000
 const makeMoreOrdinals = (
