+++ shared/chat/conversation/list-area/jump-to-recent.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
-import * as Kb from '../../../../common-adapters'
-import * as Styles from '../../../../styles'
+import * as Kb from '../../../common-adapters'
+import * as Styles from '../../../styles'
 
 type Props = {
   onClick: () => void
