+++ shared/chat/index.stories.tsx
@@ -16,7 +16,7 @@ import replyPreview from './conversation/reply-preview/index.stories'
 import setExplodingMessage from './conversation/messages/set-explode-popup/index.stories'
 import suggestors from './conversation/input-area/suggestors/index.stories'
 import messages from './conversation/messages/index.stories'
-import thread from './conversation/list-area/normal/index.stories'
+import thread from './conversation/list-area/index.stories'
 import threadSearch from './conversation/search/index.stories'
 import typing from './conversation/input-area/normal/typing/index.stories'
 import blocking from './blocking/index.stories'
